C++ uint16_t to int for ATN.

+++ runtime/Cpp/runtime/src/atn/ATNDeserializer.h
@@ -20,7 +20,7 @@ namespace atn {
 
     explicit ATNDeserializer(ATNDeserializationOptions deserializationOptions);
 
-    std::unique_ptr<ATN> deserialize(const std::vector<uint16_t> &input) const;
+    std::unique_ptr<ATN> deserialize(const std::vector<int> &input) const;
     void verifyATN(const ATN &atn) const;
 
   private:
